requirements.txt
flask==2.2.5
flask-wtf==1.2.1
selenium==4.20.0
chromedriver-autoinstaller==0.6.0
python-dotenv==1.0.1
werkzeug==2.2.3
wtforms==3.1.2
gunicorn==21.2.0
waitress==3.0.2
chrome-extension-python==1.0.1
requests==2.32.3
Pillow==10.2.0
pyvirtualdisplay==3.0
webdriver-manager==4.0.1
google-cloud-storage==2.15.0
google-cloud-logging==3.10.0
google-cloud-secret-manager==2.19.0
pymysql==1.1.0 